requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
sympy>=1.13.0
numpy>=2.0.0
requests>=2.32.0
rapidfuzz>=3.9.0
PyYAML>=6.0.2
